Fix infinite loop on provider failure
AWS includes a transaction id in the error message returned on
ProviderError. The current logic checks for differences in the previous
status message and the current and does an update if they are the
different. Since it is always different due to the id changing, it
causes it to keep updating and reconciling.
Change here just removes the check against the error message.